requirements.txt
torch
torchvision
albumentations
numpy
Pillow
matplotlib